refactor(backends): complete the Firecracker extraction
All five drivers now live in mvm-backends. mvm-runtime::driver is
re-exports only, and mvm-backends depends on neither mvm-runtime nor
mvm-build.

Into mvm-backends::fc: the API client, VMM process lifecycle, control,
observe, guards, snapshot mechanics, fork namespace, FirecrackerIO,
FcVmFullControl, driver/fc.rs, and the legacy FirecrackerBackend.
base/config.rs moved to mvm-vmm::host — a 297-line dependency-free leaf
that was the last thing pinning those modules, so neither RuntimeVolume
nor VmSlot had to move and flake_run/boot_config/run_info stayed with the
orchestration they belong to.

Zero new traits and zero new structs, and negative on net. Three pieces
of dead weight went with the move:

- the retired raw flake launcher — four public functions, no callers, and
  run_from_build was a refusal stub rather than a launcher
- require_linux_env — fn(..) -> Result<()> { Ok(()) }, called at seven
  sites, asserting nothing
- a duplicate bind_unix_listener; mvm-vmm's is now the only copy

One judgement call for review: FirecrackerBackend::start delegated to
fc_runner(), which lives in mvm-runtime and would have inverted the
dependency. FcDriver never calls it, nothing else constructs the type,
and AnyBackend routes every real start through the runner, so it refuses
loudly instead of silently succeeding — a refusal on an unreachable path,
not a stub standing in for missing work.

Tests that assert runtime orchestration rather than VMM mechanics stayed
in mvm-runtime rather than being stubbed down into the driver crate.
