+++ crates/mvm-backends/src/driver/mod.rs
@@ -1,9 +1,11 @@
 //! Concrete [`mvm_vmm::driver::VmmDriver`] implementations.
 
+pub mod fc;
 pub mod hvf;
 pub mod libkrun;
 pub mod qemu;
 
+pub use fc::FcDriver;
 pub use hvf::HvfDriver;
 pub use libkrun::LibkrunDriver;
 pub use qemu::QemuDriver;
