+++ crates/mvm-backends/src/legacy/mod.rs
@@ -7,5 +7,6 @@
 pub mod libkrun;
 pub mod qemu;
 
+pub mod fc;
 pub mod hvf;
 pub mod hvf_bootargs;
