+++ crates/mvm-runtime/src/backend.rs
@@ -1,8 +1,8 @@
 use crate::catalog;
 use anyhow::Result;
 use mvm_core::vm_backend::{
-    BackendKind, BackendSecurityProfile, ClaimStatus, LayerCoverage, SnapshotCapability, VmBackend,
-    VmCapabilities, VmId, VmInfo, VmStartConfig, VmStatus, WarmStartOutcome,
+    BackendKind, BackendSecurityProfile, SnapshotCapability, VmBackend, VmCapabilities, VmId,
+    VmInfo, VmStartConfig, VmStatus, WarmStartOutcome,
 };
 
 // Every backend variant + the FC support modules live in this crate.
@@ -10,16 +10,15 @@ use mvm_core::vm_backend::{
 // (`config`, `shell`, `runtime_meta`) lives in `crate::base`.
 use crate::apple_container_backend::AppleContainerBackend;
 use crate::base::config::{PortMapping, VmSlot};
-use crate::base::shell::run_in_vm_stdout;
 use crate::docker_backend::DockerBackend;
 use crate::driver::FcDriver;
 use crate::image::RuntimeVolume;
+use crate::microvm;
 use crate::microvm::FlakeRunConfig;
 #[cfg(feature = "test-support")]
 use crate::mock::MockBackend;
 use crate::wasm_backend::WasmBackend;
 use crate::workload_runner::{RealBrokerRegistrar, RealEndpointSpawner, WorkloadRunner};
-use crate::{firecracker, microvm};
 use mvm_backends::driver::hvf::HvfDriver;
 use mvm_backends::driver::{LibkrunDriver, QemuDriver};
 use mvm_vmm::host::drive_file::DriveFile;
@@ -190,154 +189,6 @@ fn validate_firecracker_start_config(config: &VmStartConfig) -> Result<()> {
         );
     }
     Ok(())
-}
-
-/// Firecracker backend implementation.
-///
-/// Wraps the existing free functions in [`microvm`] and [`firecracker`]
-/// behind the [`VmBackend`] trait. This is a thin adapter — all real
-/// work is delegated to the existing implementation.
-pub struct FirecrackerBackend;
-
-impl VmBackend for FirecrackerBackend {
-    fn name(&self) -> &str {
-        "firecracker"
-    }
-
-    fn kind(&self) -> catalog::BackendKind {
-        catalog::BackendKind::Firecracker
-    }
-
-    fn capabilities(&self) -> VmCapabilities {
-        // Firecracker ships a virtio-balloon device with PATCH-able
-        // target via `/balloon`; the start path attaches it whenever
-        // `VmStartConfig::mem_initial_mib` is `Some`. Capability is
-        // advertised unconditionally so the host-side controller can
-        // discover support before deciding to plumb a workload.
-        VmCapabilities {
-            pause_resume: true,
-            snapshots: false,
-            snapshot_capability: SnapshotCapability::Unsupported,
-            standby_pool: false,
-            vsock: true,
-            tap_networking: false,
-            no_routable_guest_nic: true,
-            host_vsock_proxy: true,
-            balloon: true,
-            fs_quick_checkpoint: false,
-            ..VmCapabilities::default()
-        }
-    }
-
-    fn start(&self, config: &VmStartConfig) -> Result<VmId> {
-        fc_runner().start(config)
-    }
-
-    fn stop(&self, id: &VmId) -> Result<()> {
-        microvm::stop_vm(&id.0)
-    }
-
-    fn stop_all(&self) -> Result<()> {
-        microvm::stop_all_vms()
-    }
-
-    fn pause(&self, id: &VmId) -> Result<()> {
-        microvm::pause_vm(&id.0)
-    }
-
-    fn resume(&self, id: &VmId) -> Result<()> {
-        microvm::resume_vm(&id.0)
-    }
-
-    fn balloon_set_target(&self, id: &VmId, target_inflate_mib: u32) -> Result<()> {
-        microvm::balloon_set_target(&id.0, target_inflate_mib)
-    }
-
-    fn balloon_state(&self, id: &VmId) -> Result<mvm_core::vm_backend::BalloonState> {
-        let inflated = microvm::balloon_state(&id.0)?;
-        // FC reports the inflation amount via /balloon; the cap is
-        // tracked host-side in the VM's runtime metadata (RunInfo).
-        // List the VM to recover its declared cap.
-        let vms = microvm::list_vms()?;
-        let info = vms
-            .into_iter()
-            .find(|i| i.name.as_deref() == Some(&*id.0))
-            .ok_or_else(|| anyhow::anyhow!("balloon_state: VM '{}' not found in list", id.0))?;
-        let max_mib = info.memory;
-        Ok(mvm_core::vm_backend::BalloonState {
-            max_mib,
-            inflated_mib: inflated,
-            host_committed_mib: max_mib.saturating_sub(inflated),
-        })
-    }
-
-    fn status(&self, id: &VmId) -> Result<VmStatus> {
-        let vms = microvm::list_vms()?;
-        match vms.iter().find(|info| info.name.as_deref() == Some(&*id.0)) {
-            Some(_) => Ok(VmStatus::Running),
-            None => Ok(VmStatus::Stopped),
-        }
-    }
-
-    fn list(&self) -> Result<Vec<VmInfo>> {
-        let vms = microvm::list_vms()?;
-        Ok(vms
-            .into_iter()
-            .filter_map(|info| {
-                let name = info.name.clone()?;
-                Some(VmInfo {
-                    id: VmId(name.clone()),
-                    name,
-                    status: VmStatus::Running,
-                    guest_ip: info.guest_ip,
-                    cpus: info.cpus,
-                    memory_mib: info.memory,
-                    profile: info.profile,
-                    revision: info.revision,
-                    flake_ref: info.flake_ref,
-                    ports: Vec::new(),
-                })
-            })
-            .collect())
-    }
-
-    fn logs(&self, id: &VmId, lines: u32, hypervisor: bool) -> Result<String> {
-        let abs_vms = crate::microvm::abs_vms_dir();
-        let abs_vms = abs_vms.trim();
-        let filename = if hypervisor {
-            "firecracker.log"
-        } else {
-            "console.log"
-        };
-        let log_file = format!("{}/{}/{}", abs_vms, id.0, filename);
-        run_in_vm_stdout(&format!(
-            "tail -n {} {} 2>/dev/null || true",
-            lines, log_file
-        ))
-    }
-
-    fn is_available(&self) -> Result<bool> {
-        firecracker::is_installed()
-    }
-
-    fn install(&self) -> Result<()> {
-        firecracker::install()
-    }
-
-    fn security_profile(&self) -> BackendSecurityProfile {
-        // Tier 1: full security posture. All seven CI-enforced claims
-        // hold. Hardware isolation via KVM; verified boot via
-        // dm-verity.
-        BackendSecurityProfile {
-            claims: [ClaimStatus::Holds; 7],
-            layer_coverage: LayerCoverage::all_layers(),
-            tier: "Tier 1",
-            notes: &[
-                "Full ADR-002 — all seven CI-enforced claims hold.",
-                "Hardware isolation via KVM. Verified boot via dm-verity (W3).",
-            ],
-        }
-    }
 }
 
 /// Isolation tier of a `VmBackend`.
@@ -1018,6 +869,8 @@ fn dedup_by_owning_backend(
 #[cfg(test)]
 mod tests {
     use super::*;
+    use mvm_backends::legacy::fc::FirecrackerBackend;
+    use mvm_core::vm_backend::ClaimStatus;
 
     fn vm_named(name: &str, cpus: u32) -> VmInfo {
         VmInfo {
