+++ crates/mvm-runtime/src/driver/mod.rs
@@ -5,16 +5,17 @@
 //! The concrete driver implementations now live in `mvm-backends`; this module
 //! is a compatibility re-export layer while downstream callers migrate.
 
-pub mod fc;
+#[cfg(test)]
+mod fc_orchestration_tests;
 pub mod spec;
 pub mod traits;
 
-pub use fc::FcDriver;
+pub use mvm_backends::driver::FcDriver;
 pub use mvm_backends::driver::{HvfDriver, LibkrunDriver, QemuDriver};
 // The modules as well as the types: callers reach in for driver-local helpers
 // (`driver::libkrun::map_kernel_for_test`), so re-exporting only the driver
 // types leaves those paths unresolvable.
-pub use mvm_backends::driver::{hvf, libkrun, qemu};
+pub use mvm_backends::driver::{fc, hvf, libkrun, qemu};
 pub use spec::{
     BlockDev, ConsoleCapture, KernelImage, VirtioFsShare, VmmSpec, VsockDirection, VsockPort,
 };
